fix fin extension usage
