+++ smack-experimental/src/test/java/org/jivesoftware/smackx/mam/ResultsLimitTest.java
@@ -19,10 +19,8 @@ package org.jivesoftware.smackx.mam;
 import java.lang.reflect.Method;
 
 import org.jivesoftware.smack.packet.IQ;
-import org.jivesoftware.smack.util.PacketParserUtils;
 import org.jivesoftware.smackx.mam.element.MamElements;
 import org.jivesoftware.smackx.mam.element.MamQueryIQ;
-import org.jivesoftware.smackx.mam.element.MamElements.MamFinExtension;
 import org.jivesoftware.smackx.xdata.packet.DataForm;
 import org.junit.Assert;
 import org.junit.Test;
@@ -37,12 +35,6 @@ public class ResultsLimitTest extends MamTest {
                 + "</set>" + "</query>" + "</iq>";
     }
 
-    private String getIQLimitedResultsExample() {
-        return "<iq type='result' id='u29303'>" + "<fin xmlns='urn:xmpp:mam:1' complete='true'>"
-                + "<set xmlns='http://jabber.org/protocol/rsm'>" + "<first index='0'>23452-4534-1</first>"
-                + "<last>390-2342-22</last>" + "<count>16</count>" + "</set>" + "</fin>" + "</iq>";
-    }
-
     @Test
     public void checkResultsLimit() throws Exception {
         Method methodAddResultsLimit = MamManager.class.getDeclaredMethod("addResultsLimit", Integer.class,
@@ -55,16 +47,6 @@ public class ResultsLimitTest extends MamTest {
 
         methodAddResultsLimit.invoke(mamManager, 10, mamQueryIQ);
         Assert.assertEquals(mamQueryIQ.toString(), getResultsLimitStanza(mamQueryIQ.getStanzaId(), 10));
-    }
-
-    @Test
-    public void checkQueryLimitedResults() throws Exception {
-        IQ iq = (IQ) PacketParserUtils.parseStanza(getIQLimitedResultsExample());
-
-        MamFinExtension mamFinExtension = MamFinExtension.from(iq);
-
-        Assert.assertEquals(mamFinExtension.getRSMSet().getCount(), 16);
-
     }
 
 }
